An initial approach to the Tree API, with possible issues.
