+++ pyconll/__init__.py
@@ -3,7 +3,7 @@ A library whose purpose is to provide a low level layer between the CoNLL format
 and python code.
 """
 
-__all__ = ['exception', 'load', 'tree', 'unit', 'util']
+__all__ = ['exception', 'load', 'sentencetree', 'tree', 'unit', 'util']
 
 from .load import load_from_string, load_from_file, load_from_url, \
     iter_from_string, iter_from_file, iter_from_url
