+++ pyconll/tree/__init__.py
@@ -1,9 +1,5 @@
 """
-Defines the utilties for interfacing with CoNLL sentences as trees. This
-is a utility method which when provided a Sentence constructs the appropriate 
+Defines the modules for interfacing with CoNLL sentences as trees. This
+is a utility module which when provided a Sentence constructs the appropriate 
 or corresponding tree structure.
 """
-
-__all__ = ['sentencetree']
-
-from .sentencetree import SentenceTree
