+++ pyconll/tree/_treebuilder.py
@@ -1,8 +1,10 @@
 """
-A general tree module. This module includes a Builder to create the actual Tree
-and the immutable Tree definition.
+Defines a TreeBuilder module that will actually create a Tree. This class is
+essentially a friend class of the Tree, so that the API of Tree can remain
+immutable while allowing efficient creation of a new Tree in an easy fluent way.
 """
 
+from pyconll.tree.tree import Tree
 
 class TreeBuilder:
     """
@@ -151,62 +153,3 @@ class TreeBuilder:
         self.current = new_current
 
 
-class Tree:
-    """
-    A tree node. This is the base representation for a tree, which can have many
-    children which are accessible via child index. The tree's structure is
-    immutable, so the data, parent, children cannot be changed once created.
-    """
-
-    def __init__(self):
-        """
-        Create a new empty tree. To create a useful Tree, use TreeBuilder.
-        """
-        self._data = None
-        self._parent = None
-        self._children = []
-
-    @property
-    def data(self):
-        """
-        The data on the tree node. The property ensures it is readonly.
-
-        Returns:
-            The data stored on the Tree. No data is represented as None.
-        """
-        return self._data
-
-    @property
-    def parent(self):
-        """
-        Provides the parent of the Tree. The property ensures it is readonly.
-
-        Returns:
-            A pointer to the parent Tree reference. None if there is no parent.
-        """
-        return self._parent
-
-    def __getitem__(self, key):
-        """
-        Get specific children from the Tree. This can be an integer or slice.
-
-        Args:
-            key: The indexer for the item.
-        """
-        return self._children[key]
-
-    def __iter__(self):
-        """
-        Provides an iterator over the children.
-        """
-        for child in self._children:
-            yield child
-
-    def __len__(self):
-        """
-        Provides the number of direct children on the tree.
-
-        Returns:
-            The number of direct children on the tree.
-        """
-        return len(self._children)
